refactor(search applications): rename engine name logic file
