+++ x-pack/plugins/enterprise_search/public/applications/applications/components/engine/search_application_name_logic.ts
